requirements.txt
Pillow
eth-utils